Refactor after Simons review. Mostly renaming functions
