--- src/models/UserLogin.ts
@@ -1,4 +1,0 @@
-export class UserLogin {
-	email: string;
-	password: string;
-}
