+++ src/models/auth.types.ts
@@ -1,4 +1,7 @@
-import { UserLogin } from './UserLogin';
+export class UserLogin {
+	email: string;
+	password: string;
+}
 
 export class UserCreation extends UserLogin {
 	username: string;
